separated models into their own files, and moved db to extensions file
