+++ .env-example
@@ -7,5 +7,5 @@ SECRET_KEY = ""
 STATIC_FOLDER = ""
 TEMPLATES_FOLDER = ""
 
-DATABASE_URL = ""
+DATABASE_URL = "postgresql+psycopg2://username:password@localhost:5432/mydatabase"
 SQLALCHEMY_TRACK_MODIFICATIONS = ""
