+++ reviveme/api/v1/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint("v1", __name__)
+
+from . import routes
