+++ reviveme/api/v1/routes.py
@@ -1,6 +1,7 @@
+from reviveme import db
+from reviveme.models import Thread
+
 from . import bp
-from reviveme_server import db
-from reviveme_server.models.models import *
 
 
 @bp.route("/")
@@ -8,8 +9,9 @@ from reviveme_server.models.models import *
 def index():
     return "Hello, World!"
 
+
 @bp.route("/threads", methods=["GET"])
 def thread_list():
     # TODO: figure out how to serialize our models properly
     threads = db.session.execute(db.select(Thread)).scalars().all()
-    return [ {'title': thread.title, 'content': thread.content} for thread in threads]
+    return [{"title": thread.title, "content": thread.content} for thread in threads]
