+++ reviveme/extensions.py
@@ -1,4 +1,9 @@
+from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.orm import DeclarativeBase
+
 
 class BaseModel(DeclarativeBase):
     pass
+
+
+db = SQLAlchemy(model_class=BaseModel)
