+++ reviveme/models/__init__.py
@@ -1,0 +1,3 @@
+from reviveme.models.comment import Comment
+from reviveme.models.thread import Thread
+from reviveme.models.user import User
