+++ reviveme/models/comment.py
@@ -1,0 +1,17 @@
+from sqlalchemy import ForeignKey, Text
+from sqlalchemy.orm import Mapped, mapped_column, relationship
+
+from reviveme.extensions import db
+
+
+class Comment(db.Model):
+    __tablename__ = "comments"
+    id: Mapped[int] = mapped_column(primary_key=True)
+    author_id: Mapped[int] = mapped_column(ForeignKey("users.id"))
+    thread_id: Mapped[int] = mapped_column(ForeignKey("threads.id"))
+    content: Mapped[str] = mapped_column(Text)
+
+    thread: Mapped["Thread"] = relationship("Thread", back_populates="comments")
+
+    def __repr__(self):
+        return f"<Comment id={self.id!r}, author_id={self.author_id!r}, thread_id={self.thread_id!r} content={self.content!r}>"
