--- reviveme_server/api/v1/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-bp = Blueprint('v1', __name__)
-
-from . import routes
